+++ CHANGELOG.md
@@ -10,17 +10,18 @@ Change Log -- drives
   - Converted project from a fixed Visual Studio build to use CMake instead, assuming `build/` as
     the build directory.
   - Lots of code refactoring and rewrites.
+  - Print drive GUID in human output instead of verbose output.
 
 ## Added
   - New JSON output option: `--json`.
+  - Report drive capacity and usage in verbose output.
 
 
 ----------------------------------------------------------------------------------------------------
 # v2.0.0  (2018-06-23)
 
 ## Removed
-  - No more support for `/p`, `/v`, or `/h`. Use `-p`, `-v` or `-h` only.
-    `/?` support remains.
+  - No more support for `/p`, `/v`, or `/h`. Use `-p`, `-v` or `-h` only. `/?` support remains.
 
 ## Changed
   - Converted to MIT license